+++ Release-Notes.md
@@ -1,10 +1,14 @@
 # Invoke-Build Release Notes
 
+## v5.6.3
+
+Fixed potential issues with tasks and parameters named `Count`, `Keys`, `Values`, [#173](https://github.com/nightroman/Invoke-Build/issues/173).
+
 Created [Invoke-Build.template](https://github.com/nightroman/Invoke-Build.template)
 
 ## v5.6.2
 
-Fail on adding tasks after loading build scripts, [#171](https://github.com/nightroman/Invoke-Build/issues/171).
+Fail on adding tasks after loading build scripts, #171.
 
 ## v5.6.1